Configured data model to use ERA5
Added ERA5 data to force ELM in I and IG cases.
There are also a few updates to the firn model that
include a new cold start initialization to use with
a thicker (30 m w.e.) firnpack.

All mods are intended to be BFB wrt the default configuration,
because they are wrapped within the use_extrasnowlayers
control logic. ERA5 is a new forcing dataset for I and IG cases that
only affects results when selected in place of default forcing datasets.